created separate folders for each component and page
