+++ client/src/components/Footer/index.jsx
@@ -1,4 +1,4 @@
-import Navigation from "./Navigation";
+import Navigation from "../Navigation";
 
 export default function Footer() {
     return (
