+++ client/src/main.jsx
@@ -4,10 +4,10 @@ import { ChakraProvider } from '@chakra-ui/react'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
 import App from './App';
-import Home from './pages/Home';
-import Profile from './pages/Profile';
-import Friends from './pages/Friends';
-import Grouplist from './pages/Grouplist';
+import Home from './pages/Home/index.jsx';
+import Profile from './pages/Profile/index.jsx';
+import Friends from './pages/Friends/index.jsx';
+import Grouplist from './pages/GroupList/index.jsx';
