+++ client/src/pages/Friends/index.jsx
